requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
psycopg==3.2.6
psycopg-binary==3.2.6
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
resend==2.7.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
